requirements.txt
pyspark==3.4.0
pymongo
pandas
numpy
tensorflow
scikit-learn
requests
kafka-python==2.0.2
kafka-python
ipython